Chore: Hooks app/views/DefaultBrowserView (#4424)
* Chore: Hooks app/views/DefaultBrowserView

* fix type

* minor tweak

* item on new file

* minor tweaks

* minor tweak
